[FIX] pos_category_multi: Addition of a new validation considering the notes at the moment of filtering new printing changes.
An error was happening during the use of the PoS while trying to print the orders, this error was that some products that had different
internal notes wouldn't show in the printed receipt. The solution was to add a validation inside the generation of the printing changes,
this validation now considers the notes of each product as well.